Common: Move cpu_detect to x64 directory.
